Add Meal Frequency Use Case, Change Data type of calories, quantity t…

+++ app/src/main/java/vn/edu/tdtu/anhminh/myapplication/Data/Local/DAO/IngredientDAO.java
@@ -47,7 +47,7 @@ public interface IngredientDAO {
     LiveData<List<IngredientEntity>> getQuantityAndUnitForRecipe(int recipeId);
 
     @Query("UPDATE Ingredient SET quantity = :newQuantity WHERE ingredient_id = :ingredientId")
-    int updateIngredientQuantity(int ingredientId, int newQuantity);
+    int updateIngredientQuantity(int ingredientId, double newQuantity);
 
     @Query("UPDATE Ingredient SET unit = :newUnit WHERE ingredient_id = :ingredientId")
     int updateIngredientUnit(int ingredientId, String newUnit);
